+++ R/hs.R
@@ -1,12 +1,12 @@
 #' Nonparametric calculation of univariate Value at Risk and Expected Shortfall
 #'
-#' Computes Value at Risk and Expected Shortfall (also called Conditional Value
+#' Computes Value at Risk and Expected Shortfall (Conditional Value
 #' at Risk) by means of plain and age-weighted historical simulation.
 #'
 #' @param x a numeric vector of asset returns
-#' @param p confidence level for VaR calculation; default is 0.975
-#' @param method method to be used for calculation; default is 'plain'
-#' @param lambda decay factor for the calculation of weights; default is 0.98
+#' @param p confidence level for VaR calculation; default is \code{0.975}
+#' @param method method to be used for calculation; default is \code{'plain'}
+#' @param lambda decay factor for the calculation of weights; default is \code{0.98}
 #'
 #' @export
 #'
@@ -41,7 +41,7 @@ hs <- function(x, p = 0.975, method = c("age", "plain"), lambda = 0.98) {
 
     if (method == "plain") {
         l <- sort(-x)
-        VaR <- stats::quantile(l, p)
+        VaR <- unname(stats::quantile(l, p))
         ES <- mean(l[l > VaR])
     }
 
@@ -62,9 +62,6 @@ hs <- function(x, p = 0.975, method = c("age", "plain"), lambda = 0.98) {
         l.ES <- l[l > VaR]
         ES <- sum(l.ES * w.ES)
     }
-    results <- cbind(VaR = VaR, ES = ES)
-    colnames(results) <- c("VaR", "ES")
-    rownames(results) <- paste0(100 * p, "%")
-    results <- list(VaR_ES = results)
+    results <- list(VaR = VaR, ES = ES)
     results
 }